requirements.txt
Flask==3.0.2
Python-dotenv==1.0.1